requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-import-export==4.3.7
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
phonenumbers==9.0.3
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.5
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.1
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
